refactor: #39 사용자 권한 추가

+++ coupon/src/main/java/com/tweaty/coupon/presentation/CouponController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -17,11 +18,14 @@ import com.tweaty.coupon.application.CouponService;
 import com.tweaty.coupon.application.dto.CouponCreateResponse;
 import com.tweaty.coupon.application.dto.CouponIssueResponse;
 import com.tweaty.coupon.application.dto.CouponReadResponse;
+import com.tweaty.coupon.application.dto.CouponStatusUpdateResponse;
 import com.tweaty.coupon.application.dto.CouponUpdateResponse;
 import com.tweaty.coupon.presentation.request.CouponCreateRequest;
 import com.tweaty.coupon.presentation.request.CouponIssueRequest;
 import com.tweaty.coupon.presentation.request.CouponUpdateRequest;
 
+import constant.UserConstant;
+import domain.Role;
 import lombok.RequiredArgsConstructor;
 
 @RestController
@@ -31,7 +35,12 @@ public class CouponController {
 	private final CouponService couponService;
 
 	@PostMapping()
-	public ResponseEntity<CouponCreateResponse> createCoupon(@RequestBody CouponCreateRequest request) {
+	public ResponseEntity<CouponCreateResponse> createCoupon(
+		@RequestBody CouponCreateRequest request,
+		@RequestHeader(UserConstant.X_USER_ROLE) Role role
+	) {
+		Role.checkAdmin(role);
+
 		CouponCreateResponse response = couponService.createCoupon(request);
 		return ResponseEntity.status(HttpStatus.CREATED).body(response);
 	}
@@ -39,11 +48,10 @@ public class CouponController {
 	@PostMapping("/{couponId}/issue")
 	public ResponseEntity<CouponIssueResponse> issueCoupon(
 		@PathVariable UUID couponId,
+		@RequestHeader(UserConstant.X_USER_ID) UUID customerId,
 		@RequestBody CouponIssueRequest request
 	) {
-		UUID customerId = UUID.randomUUID();
-		//CouponIssueResponse response = couponService.issueCoupon(couponId, customerId, request);
-		CouponIssueResponse response = couponService.issueCoupon(couponId, request);
+		CouponIssueResponse response = couponService.issueCoupon(couponId, customerId, request);
 		return ResponseEntity.status(HttpStatus.CREATED).body(response);
 	}
 
@@ -56,15 +64,41 @@ public class CouponController {
 	@PatchMapping("/{couponId}")
 	public ResponseEntity<CouponUpdateResponse> updateCoupon(
 		@PathVariable UUID couponId,
+		@RequestHeader(UserConstant.X_USER_ROLE) Role role,
 		@RequestBody CouponUpdateRequest request
 	) {
+		Role.checkAdmin(role);
+
 		CouponUpdateResponse response = couponService.updateCoupon(couponId, request);
 		return ResponseEntity.status(HttpStatus.ACCEPTED).body(response);
 	}
 
 	@DeleteMapping("/{couponId}")
-	public ResponseEntity<Void> deleteCoupon(@PathVariable UUID couponId) {
+	public ResponseEntity<Void> deleteCoupon(
+		@PathVariable UUID couponId,
+		@RequestHeader(UserConstant.X_USER_ROLE) Role role
+	) {
+		Role.checkAdmin(role);
+
 		couponService.deleteCoupon(couponId);
 		return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+	}
+
+	@PatchMapping("/{couponId}/use")
+	public ResponseEntity<CouponStatusUpdateResponse> useCoupon(
+		@PathVariable UUID couponId,
+		@RequestHeader(UserConstant.X_USER_ID) UUID customerId
+	) {
+		CouponStatusUpdateResponse response = couponService.useCoupon(couponId, customerId);
+		return ResponseEntity.status(HttpStatus.ACCEPTED).body(response);
+	}
+
+	@PatchMapping("/{couponId}/cancel")
+	public ResponseEntity<CouponStatusUpdateResponse> cancelCoupon(
+		@PathVariable UUID couponId,
+		@RequestHeader(UserConstant.X_USER_ID) UUID customerId
+	) {
+		CouponStatusUpdateResponse response = couponService.cancelCoupon(couponId, customerId);
+		return ResponseEntity.status(HttpStatus.ACCEPTED).body(response);
 	}
 }
